+++ petname-macros/src/lang/mod.rs
@@ -1,0 +1,4 @@
+//! Language-specific petname macros.
+
+pub(crate) mod english;
+pub(crate) mod turkish;
